// src/modules/storage/enums/storage-file-type.enum.ts
export enum StorageFileType {
  PLAN_TRABAJO = "PLAN_TRABAJO",
  INFORME_SEMESTRAL = "INFORME_SEMESTRAL",
  INFORME_ADMISION = "INFORME_ADMISION",
  ACTAS = "ACTAS",
  REPORTE_MENSUAL = "REPORTE_MENSUAL",
  SEGUIMIENTO_ACOMPANANTE = "SEGUIMIENTO_ACOMPANANTE",
  SEGUIMIENTO_ACOMPANIANTE_EXTERNO = "SEGUIMIENTO_ACOMPANIANTE_EXTERNO",
  SEGUIMIENTO_FAMILIA = "SEGUIMIENTO_FAMILIA",
  FACTURA = "FACTURA",
  FIRMA_DIGITAL = "FIRMA_DIGITAL",
}